[rt] Update to 3.2.26-rt39
This is a bit of a kluge, as Steven Rostedt is no longer providing
rebased tags in git.  There is one updated (23) and one new patch
(282) from 3.2.24-rt38-rebase, and the version bump patch that follows
the trivial merge of 3.2.26 and 3.2.24-rt38 (283).

The combined diff for this patch series appears to be the same (modulo
line numbers shifted by Debian patches) as that between 3.2.26 and
3.2.26-rt39.


git-svn-id: svn://anonscm.debian.org/svn/kernel/dists/sid/linux/debian@19313 510b9475-24dd-0310-9b6c-e0eefe99d49f