linux-source: Install kernel config files for linux-image packages
The kernel config files used to be uploaded to
http://kernel.alioth.debian.org/config/ or elsewhere, and people
occasionally request them, but we have no infrastructure to
automatically do that and I rarely remember to do it.  These are
presumably wanted as a starting point by people building custom
packages who don't want to download the linux-image package, so
add them to the linux-source-$UPSTREAMVERSION package.

We can already generate all the config files without any
cross-toolchain using the 'setup' target in rules.gen, except that the
'make prepare' command in rules.real will fail.  This command is
unnecessary as 'prepare' is a dependency of any build target we use
in our 'build-arch' rule, so remove it.


git-svn-id: svn://anonscm.debian.org/svn/kernel/dists/sid/linux/debian@20400 510b9475-24dd-0310-9b6c-e0eefe99d49f