aufs: Update to aufs3.x-rcN-20131104
mmap() holds reference to the union file as well as the branch file.
This behaviour requires an extra patch and used to be conditional on
CONFIG_AUFS_PROC_MAP.  It is now unconditional because otherwise the
aufs inode number may be reused and this can break userland file
caching.  So add the extra patch and drop our patch to disable
CONFIG_AUFS_PROC_MAP.


git-svn-id: svn://anonscm.debian.org/svn/kernel/dists/trunk/linux/debian@20778 510b9475-24dd-0310-9b6c-e0eefe99d49f