Remove non i386/amd64 arches, rt featureset, patches, abi
This will be an x86-only kernel package, so remove non-x86 arches.

This is for i-pipe-based kernels, which don't match up with
RT_PREEMPT, so remove the rt featureset.

This should be close to vanilla in order for Xenomai dev support, so
remove distribution patches.

This will be updated, so remove 3.2.0-4 abi files.