Integrate r8168 NIC driver
The upstream kernel and Debian packages don't include the r8168
driver, and the r8169 driver is autoloaded when an r8168 chipset is
detected.  The r8169 driver is notoriously unreliable when driving the
r8168, and only drives it at 100Mb when it does work.

For unknown reasons, the r8169 driver stops working altogether in
otherwise-identical kernels to which the Xenomai patches have been
applied, so this kernel includes the r8168 driver, blacklisted by
default.

Downloaded from:

http://r8168dl.appspot.com/files/r8168-8.038.00.tar.bz2

Changes:

- Include the driver files as a patch
- Plumb the driver source into kbuild
- Add the module to the package installer files
- Generate a modprobe.d blacklist file from the postinst script *

* The r8168 driver conflicts with the r8169 driver.  Kernel packages
must not contain conflicting files, so all filenames must be unique.
/etc/modprobe.d/r816x_blacklist.conf can't be included as a packaged
file, so instead it's created by the postinst script, with the r8168
driver blacklisted by default.

Conflicts:
	changelog
	config/config
	patches/series