Add xenomai.beaglebone featureset (stub)
This is a stub for the Beaglebone.  It needs patches and kernel
configs.

*** Previous version:

Previous version Cribbed from Robert Nelson's packaging:

    https://github.com/RobertCNelson/bb-kernel/tree/am33x-v3.8/patches/packaging

Import RCN patches and config for xenomai.beaglebone featureset

Patches and config added from git rev a8e5933000, 2014-05-20:

    https://github.com/cdsteinkuehler/linux-dev

- omap flavour featureset configuration
- base patch series template
- featureset rules file

The 'firmware/0001-firmware-add-for-beaglebone.patch' is a binary
patch, not understood by quilt.  This was added with the
'include-binaries' Debian packaging feature, and an install rule added
to rules.featureset-xenomai.beaglebone.

Add drivers to RCN beaglebone config:
- Ralink RT2X00
- autofs4

- get rid of upstream omap, mx5, vexpress config snippets:  not applicable
  - omap snippet causing problems when applied on top of .rcn?
- empty out armhf/config
- replace config.omap with config.rcn
- fix '[image]/configs:' in defines files