Update x32 syscall patch to block system calls >= 512 by default
System calls from x32 tasks are distinguished by having bit 30 set,
but they share the system call table with x86_64 so where parameter/
return value adjustment is needed there is a difference in the low
bits too.  The x32-specific calls are numbered from 512 and of course
are not present in the table if the kernel doesn't support x32.

This means we need to change both the maximum syscall number and the
mask instruction.


git-svn-id: svn://anonscm.debian.org/svn/kernel/dists/sid/linux/debian@21689 510b9475-24dd-0310-9b6c-e0eefe99d49f